requirements.txt
cryptography
Flask
gunicorn
pillow
pymysql